langchain_project/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.5
click==8.1.6
coloredlogs==15.0.1
comm==0.1.3
dataclasses-json==0.5.14
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.1
docx2txt==0.8
emoji==2.8.0
executing==1.2.0
fastapi==0.99.1
fastjsonschema==2.18.0
filetype==1.2.0
flatbuffers==23.5.26
fqdn==1.5.1
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
langchain==0.0.251
langsmith==0.0.18
loguru==0.7.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
nltk==3.8.1
notebook==7.0.1
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.25.2
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.16.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
platformdirs==3.10.0
posthog==3.0.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pypdf==3.14.0
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
streamlit==1.25.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
terminado==0.17.1
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.10.2
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
validators==0.20.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
widgetsnbextension==4.0.8
wikipedia==1.4.0
yarl==1.9.2
zipp==3.16.2
